mlpwr engine

mlpwr engine

Usage

calculate_mlpwr(
  test_n,
  n_reps_total,
  n_reps_per,
  se_final,
  min_sample_size,
  max_sample_size,
  target_performance,
  c_statistic,
  mean_or_assurance,
  n_init,
  verbose,
  data_function,
  model_function,
  metric_function,
  value_on_error
)

Arguments

test_n

The sample size used for test datasets. This should be a large number.

n_reps_total

The total number of simulation reps run by pmsims.

n_reps_per

The number of reps run at each sample size.

se_final

A standard error which can be used as a stopping criteria. Either n_reps_total or se_final should be given.

min_sample_size

The minimum sample size assessed. This sets the lower bound of the search region for sample size.

max_sample_size

The maximum sample size assessed. This sets the upper bound of the search region for sample size.

target_performance

The minimum desired model performance

mean_or_assurance

Can be either "mean" or "assurance". If mean, sample size is calculated so that average model performance is greater than target_performance. If assurance model performance is greater than target performance 80% of the time.

n_init

The number of initial sample sizes simualted before the Gausian process search begins.

verbose

logical. Passed to mlpwr; return verbose output.

data_function

A function that returns datasets. Must have a single argmument, n, which controls the sample size.

model_function

A function that fits models to the data. Take the data object returned by data_funciton as only argument.

metric_function

A function that returns a performance metric. Must take test data, a fitted model and a model function as arguments. Must return a single value.

value_on_error

The value used if there is an error in fitting the model or calculating performance.