Title

Title

Usage

default_tune(
  tune_param,
  max_sample_size,
  large_sample_performance,
  data_function,
  model_function,
  metric_function,
  tune_args = NULL,
  verbose = TRUE
)

Arguments

tune_param

The name of the tuning parameter in the data generating model

max_sample_size

The maximum sample size used in simulations

large_sample_performance

The target large sample performance

data_function

The data generating funciton

model_function

The estimation model function

metric_function

The performance metric function the model is to be tuned with.

tune_args

A list of arguments which may include interval, large_n, tolerance, and max_interval_expansion. These will overide defaults

Value

The optimal value for the tuning parameter