Simulate Continuous data

Simulate Continuous data

Usage

generate_continuous_data(
  n,
  beta_signal,
  n_signal_parameters,
  noise_parameters,
  predictor_type,
  predictor_prop = NULL
)

Arguments

n

Sample size

beta_signal

Association between signal predictors and the outcome

n_signal_parameters

Number of predictors that have a non zero association with the outcome @param noise_parameters Number of predictors with no association with outcome

predictor_type

Type of predictor, can be "continuous" or "binary."

predictor_prop

If predictor type is binary, the probability of a predictor taking value 1

Value

A data frame with one outcome column and n_signal_parameters + noise_parameters predictor columns

Examples

generate_continuous_data(
  n = 100, n_signal_parameters = 10,
  noise_parameters = 10, predictor_type = "binary", predictor_prop = 0.1,
  beta_signal = 0.1
)
#>                y x1 x2 x3 x4 x5 x6 x7 x8 x9 x10 x11 x12 x13 x14 x15 x16 x17 x18
#> 1   -0.136714960  0  0  0  0  0  0  0  0  1   1   0   0   0   0   0   0   1   0
#> 2   -0.115840184  0  0  1  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 3    0.721132287  0  0  0  0  1  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 4   -1.084026517  0  0  0  0  0  1  0  0  0   1   0   0   0   0   0   0   0   0
#> 5   -1.300092441  0  0  0  0  0  0  0  0  0   0   0   1   0   0   0   0   0   0
#> 6   -0.376769462  0  0  0  0  0  0  0  0  0   0   0   0   0   0   1   0   0   1
#> 7    0.203748646  0  0  0  0  0  0  0  1  0   0   0   0   1   0   0   0   0   0
#> 8   -0.703562273  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 9    1.697413940  0  0  0  0  0  1  0  1  0   0   0   0   0   0   0   0   0   0
#> 10  -0.202826878  0  0  1  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 11  -1.276831335  0  0  0  0  0  0  0  1  0   0   0   0   0   0   0   0   0   0
#> 12   0.883170175  0  0  0  0  0  0  0  0  0   0   0   0   0   1   0   0   1   0
#> 13   0.671420279  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 14   1.430022375  0  0  0  1  0  0  0  0  1   0   0   0   0   0   0   0   0   0
#> 15   2.045467879  0  1  0  0  0  0  1  1  1   0   0   0   0   0   0   0   0   1
#> 16  -0.082926039  0  0  1  0  0  0  0  0  0   0   0   1   0   0   0   0   0   0
#> 17  -1.279257815  0  0  0  0  0  0  1  0  0   0   0   1   0   0   0   0   0   0
#> 18   1.250600601  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 19  -1.676236775  0  0  0  0  1  1  0  0  0   0   0   0   0   0   0   0   0   0
#> 20  -0.168787568  1  0  0  0  1  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 21  -2.129297308  0  0  0  0  0  0  0  0  0   0   0   0   0   0   1   0   0   0
#> 22   0.337646027  0  0  0  0  0  0  0  0  0   0   0   1   0   1   0   0   0   1
#> 23  -1.574574848  0  0  0  0  0  0  0  0  0   0   1   0   0   0   0   1   1   0
#> 24   0.003680585  0  1  0  1  0  0  0  0  0   0   0   0   1   0   0   1   0   0
#> 25   1.115060108  1  0  0  0  0  0  0  0  0   1   0   0   0   0   0   0   1   0
#> 26   1.078636538  0  0  0  0  0  0  0  0  0   0   0   1   0   0   0   0   0   0
#> 27   1.552802854  0  0  0  0  0  0  0  1  0   0   0   1   0   0   0   0   0   0
#> 28  -1.353281089  0  0  1  0  1  0  0  0  1   0   0   0   0   0   0   0   0   0
#> 29  -2.661867518  0  0  0  0  0  0  0  0  1   0   0   0   0   0   0   0   0   0
#> 30   0.264369164  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 31   1.124571594  0  0  1  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 32  -0.559185548  0  0  0  0  0  0  0  0  0   1   0   0   0   0   0   0   0   0
#> 33  -0.286304379  0  0  0  0  0  1  0  0  0   0   0   0   1   0   0   0   0   0
#> 34  -0.042498949  0  1  0  0  1  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 35   0.220593240  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 36  -1.481090030  0  0  0  0  0  0  0  0  0   0   1   0   0   0   0   0   1   0
#> 37   0.724691204  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 38  -2.277514701  1  0  0  0  0  0  0  0  0   0   0   0   0   0   0   1   0   0
#> 39  -0.065409209  0  0  0  0  0  0  0  0  0   0   0   0   0   1   0   0   0   0
#> 40  -0.169704813  0  1  0  1  0  0  0  1  0   0   0   0   0   0   0   0   0   0
#> 41  -0.180516288  0  0  1  0  0  0  0  0  0   0   0   0   0   0   0   0   1   0
#> 42   0.774836585  0  1  0  0  0  0  0  0  1   0   1   0   0   0   0   0   0   0
#> 43  -0.205676846  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   1   0
#> 44   0.334880505  0  0  0  0  0  1  0  0  0   0   0   0   0   0   0   0   0   1
#> 45   0.410628321  0  1  1  0  0  0  1  0  0   0   0   0   0   0   0   0   1   0
#> 46   0.374202233  0  0  0  0  0  1  0  0  0   0   1   0   0   0   0   0   0   0
#> 47   1.574919630  0  0  0  0  0  0  0  0  0   0   0   0   0   0   1   0   0   0
#> 48   0.767767832  1  0  0  0  0  0  0  0  0   1   0   1   0   0   0   0   0   0
#> 49   0.240387750  1  1  0  0  0  0  1  0  0   0   0   0   0   0   0   0   0   0
#> 50  -0.049587715  0  0  0  0  0  0  0  1  0   0   0   0   0   0   0   0   0   0
#> 51   0.366168788  0  0  0  0  0  1  0  0  0   0   0   0   0   0   0   0   0   0
#> 52   0.528403243  0  0  0  0  0  0  0  0  0   0   1   0   0   0   0   0   0   0
#> 53  -0.479388093  0  0  0  0  0  1  0  0  0   0   0   1   0   0   0   0   0   0
#> 54   0.942199374  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   1   0   0
#> 55  -0.606361360  0  0  0  0  0  0  1  0  0   0   0   0   0   0   0   0   0   0
#> 56  -1.175402100  0  0  1  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 57   0.733259917  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 58  -0.420207094  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 59  -0.155375086  0  0  0  1  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 60  -0.169238217  0  1  0  0  0  0  0  0  1   0   0   0   0   0   0   1   0   1
#> 61   1.312448330  0  1  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 62  -0.345363006  0  0  1  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 63   1.238667521  0  1  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 64   0.390523483  0  0  0  1  0  1  0  0  0   0   0   1   0   0   0   0   0   0
#> 65  -1.366393612  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 66  -1.198102601  0  1  0  0  0  0  0  0  0   0   0   0   1   0   0   0   0   0
#> 67  -1.937098312  1  0  0  1  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 68   0.103794475  0  0  0  1  0  0  1  0  0   0   0   0   0   0   0   0   0   0
#> 69   0.794509055  0  0  0  0  0  0  0  0  0   1   0   0   0   0   1   0   0   0
#> 70  -0.474591230  0  0  0  0  0  0  0  0  0   0   0   0   1   0   0   0   0   0
#> 71   1.863130612  0  0  0  1  0  0  0  0  0   0   0   1   0   1   0   0   0   0
#> 72   2.125846064  0  0  1  0  0  0  0  0  0   0   0   0   0   1   0   0   0   1
#> 73   0.427618809  0  0  0  0  0  0  0  0  0   0   0   0   0   1   0   0   0   0
#> 74  -0.366479349  0  1  1  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 75  -0.733835311  0  0  0  1  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 76   0.398259155  0  0  0  0  0  0  0  0  0   0   0   1   0   1   1   0   0   0
#> 77  -1.274537931  0  0  0  0  0  1  0  0  0   1   0   0   0   0   0   0   1   0
#> 78  -0.295441200  0  0  0  0  0  0  0  0  1   0   0   0   0   0   0   0   0   1
#> 79  -0.130702468  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   1   0   0
#> 80  -1.799493950  0  0  0  0  0  0  1  1  0   0   0   1   0   0   0   0   0   0
#> 81  -0.995930861  0  0  0  0  0  0  0  0  0   0   1   1   0   0   0   1   0   0
#> 82   0.520113998  0  0  0  0  1  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 83   0.137138371  0  0  0  0  1  0  1  0  0   0   0   0   0   0   0   0   0   0
#> 84  -0.080131848  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 85   0.167716586  0  0  1  0  0  0  0  1  0   0   1   0   0   0   0   0   0   0
#> 86  -0.518980925  1  0  0  1  0  0  0  0  0   0   0   0   0   0   0   1   0   0
#> 87  -1.116561323  0  0  0  0  0  0  0  0  0   0   0   0   0   0   1   0   0   0
#> 88  -0.780269903  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 89  -1.576958529  0  0  0  1  0  1  0  0  0   0   0   0   0   0   0   1   0   0
#> 90  -0.427834874  0  0  0  0  0  0  0  0  0   0   0   0   0   0   1   0   1   0
#> 91  -1.431027012  1  1  1  0  1  0  1  1  0   0   0   0   0   0   0   0   0   0
#> 92   2.750764753  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 93   1.513669721  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   1   0   0
#> 94   0.033988937  0  0  0  0  0  0  0  0  0   0   0   0   1   0   0   0   0   0
#> 95   0.740792784  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0
#> 96  -1.070808180  0  0  0  0  0  0  0  1  0   1   0   0   0   0   0   1   0   1
#> 97   0.036435768  0  1  0  0  0  0  0  0  1   0   0   0   0   0   0   1   0   0
#> 98  -0.610945489  0  0  0  0  0  0  0  0  0   0   1   0   0   0   0   0   0   1
#> 99   1.551884192  0  0  1  0  0  1  0  0  0   1   0   0   0   0   0   0   0   0
#> 100  0.450319466  0  1  0  0  0  0  0  0  0   1   0   0   0   0   0   0   1   0
#>     x19 x20
#> 1     1   0
#> 2     0   0
#> 3     1   0
#> 4     0   0
#> 5     0   1
#> 6     0   1
#> 7     0   1
#> 8     0   0
#> 9     0   0
#> 10    0   0
#> 11    0   0
#> 12    0   0
#> 13    0   1
#> 14    1   1
#> 15    0   0
#> 16    0   0
#> 17    0   0
#> 18    0   0
#> 19    0   0
#> 20    0   0
#> 21    0   0
#> 22    0   0
#> 23    0   0
#> 24    0   0
#> 25    0   1
#> 26    0   0
#> 27    0   0
#> 28    0   0
#> 29    0   0
#> 30    0   0
#> 31    0   0
#> 32    0   0
#> 33    0   0
#> 34    0   0
#> 35    0   0
#> 36    0   0
#> 37    0   0
#> 38    1   0
#> 39    0   0
#> 40    0   0
#> 41    0   0
#> 42    1   0
#> 43    1   0
#> 44    0   0
#> 45    0   0
#> 46    0   1
#> 47    1   0
#> 48    0   0
#> 49    0   0
#> 50    0   0
#> 51    0   0
#> 52    0   0
#> 53    0   0
#> 54    0   0
#> 55    0   0
#> 56    0   0
#> 57    0   1
#> 58    0   0
#> 59    0   0
#> 60    0   0
#> 61    0   0
#> 62    1   0
#> 63    0   0
#> 64    0   0
#> 65    0   0
#> 66    0   0
#> 67    0   0
#> 68    0   0
#> 69    1   1
#> 70    0   0
#> 71    0   0
#> 72    1   0
#> 73    0   0
#> 74    1   0
#> 75    0   0
#> 76    0   0
#> 77    0   0
#> 78    0   0
#> 79    0   1
#> 80    0   0
#> 81    0   0
#> 82    0   1
#> 83    0   0
#> 84    0   0
#> 85    0   0
#> 86    0   0
#> 87    0   0
#> 88    0   0
#> 89    0   0
#> 90    0   0
#> 91    0   0
#> 92    0   0
#> 93    1   0
#> 94    0   0
#> 95    0   0
#> 96    0   0
#> 97    0   0
#> 98    1   0
#> 99    0   0
#> 100   0   0