Title Tune generate data

Title Tune generate data

Usage

tune_generate_data(
  data_function,
  large_n,
  interval,
  model_function,
  metric_function,
  target_performance,
  tolerance = set_tolerance(target_performance),
  max_interval_expansion = 10,
  verbose = FALSE
)

Arguments

data_function

A function of two parameters, n and a tuning parameter, that returns data for the model function

large_n

A large sample size used for parameter tuning

interval

A vector containing the end-points of the interval to be searched

model_function

A function which takes the object returned by the data generating function and fits the analysis model of interest.

target_performance

The desired model performance in a large sample

tolerance

The tolerance in the large sample performance

max_interval_expansion

The maximum number of time the search interval will be expanded before tune generate data quits looking. This prevents getting stuck in impossible searches.

performance_function

A function which takes a a test dataset and model object as arguments and returns a performance metric

Value

optimal value for second argument of the data generating function