Calculate the minimum sample size required to develop a prediction model

Minimum working example using the mlpwr package

Usage

simulate_binary(
  n_signal_parameters,
  baseline_prob,
  min_sample_size,
  max_sample_size,
  large_sample_performance,
  minimum_threshold = 0.1,
  noise_parameters = 0,
  predictor_type = "continuous",
  predictor_prop = NULL,
  metric = "auc",
  model = "glm",
  se_final = 0.005,
  n_reps_total = NULL,
  tune_param = NULL,
  ...
)

Arguments

n_signal_parameters

Number of predictors that have a non zero association with the outcome @param noise_parameters Number of predictors with no association with outcome

baseline_prob

Baseline probability of outcome (ie. probability when all predicors are 0)

min_sample_size

The minimum sample size used in simulations

max_sample_size

The maximum sample size used in simulations

large_sample_performance

The desired performance in a large sample (for the metric defined by tune_metric_function). This may be specified in place of tune_param. The data generating model is tuned so the desired performance is obtained when n is equal to the max_sample_size.

predictor_type

Type of predictor, can be "continuous" or "binary."

predictor_prop

If predictor type is binary, the probability of a predictor taking value 1

tune_param

A tuning parameter to be passed to the data generating function

...

Other options passed to simulate_custom()

data_generating_function

A function of two parameters, n and a tuning parameter, that returns data for the model function

model_function

A function which takes the object returned by the data generating function and fits the analysis model of interest.

metric_function

A function which takes a a test dataset and model object as arguments and returns a performance metricß

target_performance

The desired performance of the prediction model

test_n

The sample size used for testing model performance

tune_args

A named list of arguments to be passed to tune_generate_data.R. Possible arguments are large_n, min_tune_arg, max_tune_arg, max_interval_expansion, and tolerance. See tune_generate_data for more details.

tune_metric_function

The metric_function used when tuning the model. This may differ from the metric function used for determining the sample size.

n_reps_totalThe

number of simulation reps

final_estimate_se

The standard error for the estimate of model performance at the minimum sample size. Either this or nreps should be specified.

n_sample_sizes

The number of different sample sizes simulations are carried out at each sample size

n_init

The number of different sample sizes for initialization (before updates)

Value

A list of results form the simulation

Examples