Calculate the minimum sample size required for a survival outcome

Calculate the minimum sample size required for a survival outcome

Usage

simulate_survival(
  n_signal_parameters,
  min_sample_size,
  max_sample_size,
  noise_parameters = 0,
  predictor_type = "continuous",
  predictor_prop = NULL,
  baseline_hazard = 0.01,
  censoring_rate = 0.2,
  metric = "cindex",
  model = "coxph",
  large_sample_performance = 0.8,
  minimum_threshold = 0.1,
  se_final = 0.005,
  n_reps_total = NULL,
  ...
)

Arguments

n_signal_parameters

Number of predictors that have a non zero association with the outcome @param noise_parameters Number of predictors with no association with outcome

predictor_type

Type of predictor, can be "continuous" or "binary."

predictor_prop

If predictor type is binary, the probability of a predictor taking value 1

baseline_hazard

Baseline Hazard

censoring_rate

Early drop out/censoring rate